refactor: directory structure for food add/edit form
